接口完成90%

+++ src/main/java/com/zerogchat/web/ChatChatController.java
@@ -3,13 +3,12 @@ package com.zerogchat.web;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
+import com.zerogchat.common.*;
 import com.zerogchat.entity.*;
-import com.zerogchat.common.ApiResult;
-import com.zerogchat.common.PageList;
-import com.zerogchat.common.ResultCode;
 import com.zerogchat.service.*;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,7 +16,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 /**
  * 控制层
@@ -32,7 +33,17 @@ public class ChatChatController {
     @Autowired
     ChatChatService service;
 
+    @Autowired
+    private ChatChatService chatService;
 
+    @Autowired
+    private ChatConfigsService configsService;
+
+    @Autowired
+    private RedisTemplate redisTemplate;
+
+    RedisHelp redisHelp =new RedisHelp();
+    ResultAll Result = new ResultAll();
     /***
      * 添加聊天室
      */
@@ -40,18 +51,28 @@ public class ChatChatController {
     @ResponseBody
     public String chatAdd(@RequestParam(value = "params", required = false) String  params,
                           @RequestParam(value = "params", required = false) String  token) {
-        ChatChat insert = null;
-        if (StringUtils.isNotBlank(params)) {
-            JSONObject object = JSON.parseObject(params);
-            insert = object.toJavaObject(ChatChat.class);
+        try {
+            ChatConfigs configs = configsService.selectByKey(0);
+            String oldToken = configs.getToken();
+            if(!oldToken.equals(token)){
+                return Result.getResultJson(0,"密钥不正确",null);
+            }
+            ChatChat insert = null;
+            if (StringUtils.isNotBlank(params)) {
+                JSONObject object = JSON.parseObject(params);
+                insert = object.toJavaObject(ChatChat.class);
+            }
+
+            int rows = service.insert(insert);
+
+            JSONObject response = new JSONObject();
+            response.put("code" , rows);
+            response.put("msg"  , rows > 0 ? "添加成功" : "添加失败");
+            return response.toString();
+        }catch (Exception e){
+            e.printStackTrace();
+            return Result.getResultJson(0,"程序执行异常，请联系管理员",null);
         }
-
-        int rows = service.insert(insert);
-
-        JSONObject response = new JSONObject();
-        response.put("code" , rows);
-        response.put("msg"  , rows > 0 ? "添加成功" : "添加失败");
-        return response.toString();
     }
 
     /***
@@ -59,19 +80,30 @@ public class ChatChatController {
      */
     @RequestMapping(value = "/chatUpdate")
     @ResponseBody
-    public String chatUpdate(@RequestParam(value = "params", required = false) String  params) {
-        ChatChat update = null;
-        if (StringUtils.isNotBlank(params)) {
-            JSONObject object = JSON.parseObject(params);
-            update = object.toJavaObject(ChatChat.class);
+    public String chatUpdate(@RequestParam(value = "params", required = false) String  params,
+                             @RequestParam(value = "params", required = false) String  token) {
+        try {
+            ChatConfigs configs = configsService.selectByKey(0);
+            String oldToken = configs.getToken();
+            if(!oldToken.equals(token)){
+                return Result.getResultJson(0,"密钥不正确",null);
+            }
+            ChatChat update = null;
+            if (StringUtils.isNotBlank(params)) {
+                JSONObject object = JSON.parseObject(params);
+                update = object.toJavaObject(ChatChat.class);
+            }
+
+            int rows = service.update(update);
+
+            JSONObject response = new JSONObject();
+            response.put("code" , rows);
+            response.put("msg"  , rows > 0 ? "修改成功" : "修改失败");
+            return response.toString();
+        }catch (Exception e){
+            e.printStackTrace();
+            return Result.getResultJson(0,"程序执行异常，请联系管理员",null);
         }
-
-        int rows = service.update(update);
-
-        JSONObject response = new JSONObject();
-        response.put("code" , rows);
-        response.put("msg"  , rows > 0 ? "修改成功" : "修改失败");
-        return response.toString();
     }
 
     /***
@@ -79,9 +111,24 @@ public class ChatChatController {
      */
     @RequestMapping(value = "/chatDelete")
     @ResponseBody
-    public int chatDelete(@RequestParam(value = "key", required = false) String  key) {
+    public String chatDelete(@RequestParam(value = "key", required = false) String  key,
+                          @RequestParam(value = "params", required = false) String  token) {
 
-        return service.delete(key);
+        try {
+            ChatConfigs configs = configsService.selectByKey(0);
+            String oldToken = configs.getToken();
+            if(!oldToken.equals(token)){
+                return Result.getResultJson(0,"密钥不正确",null);
+            }
+            int rows =  service.delete(key);
+            JSONObject response = new JSONObject();
+            response.put("code" , rows);
+            response.put("msg"  , rows > 0 ? "操作成功" : "操作失败");
+            return response.toString();
+        }catch (Exception e){
+            e.printStackTrace();
+            return Result.getResultJson(0,"程序执行异常，请联系管理员",null);
+        }
     }
 
     /***
@@ -95,18 +142,55 @@ public class ChatChatController {
     public String chatList (@RequestParam(value = "searchParams", required = false) String  searchParams,
                             @RequestParam(value = "page"        , required = false, defaultValue = "1") Integer page,
                             @RequestParam(value = "limit"       , required = false, defaultValue = "15") Integer limit) {
-        ChatChat query = new ChatChat();
-        if (StringUtils.isNotBlank(searchParams)) {
-            JSONObject object = JSON.parseObject(searchParams);
-            query = object.toJavaObject(ChatChat.class);
+        try {
+            ChatChat query = new ChatChat();
+            String sqlParams = "null";
+            if(limit>50){
+                limit = 50;
+            }
+            Integer total = 0;
+            List jsonList = new ArrayList();
+            if (StringUtils.isNotBlank(searchParams)) {
+                JSONObject object = JSON.parseObject(searchParams);
+                query = object.toJavaObject(ChatChat.class);
+                Map paramsJson = JSONObject.parseObject(JSONObject.toJSONString(query), Map.class);
+                sqlParams = paramsJson.toString();
+            }
+            total = service.total(query);
+            List cacheList = redisHelp.getList("chatList_"+page+"_"+limit+"_"+sqlParams,redisTemplate);
+            try {
+                if (cacheList.size() > 0) {
+                    jsonList = cacheList;
+                } else {
+                    PageList<ChatChat> pageList = service.selectPage(query, page, limit);
+                    jsonList = pageList.getList();
+                    if(jsonList.size() < 1){
+                        JSONObject noData = new JSONObject();
+                        noData.put("code" , 1);
+                        noData.put("msg"  , "");
+                        noData.put("data" , new ArrayList());
+                        noData.put("count", 0);
+                        return noData.toString();
+                    }
+                    redisHelp.delete("chatList_"+page+"_"+limit+"_"+sqlParams,redisTemplate);
+                    redisHelp.setList("chatList_"+page+"_"+limit+"_"+sqlParams,jsonList,3,redisTemplate);
+                }
+            }catch (Exception e){
+                e.printStackTrace();
+                if(cacheList.size()>0){
+                    jsonList = cacheList;
+                }
+            }
+            JSONObject response = new JSONObject();
+            response.put("code" , 1);
+            response.put("msg"  , "");
+            response.put("data" , jsonList);
+            response.put("count", jsonList.size());
+            response.put("total", total);
+            return response.toString();
+        }catch (Exception e){
+            e.printStackTrace();
+            return Result.getResultJson(0,"程序执行异常，请联系管理员",null);
         }
-
-        PageList<ChatChat> pageList = service.selectPage(query, page, limit);
-        JSONObject response = new JSONObject();
-        response.put("code" , 0);
-        response.put("msg"  , "");
-        response.put("data" , null != pageList.getList() ? pageList.getList() : new JSONArray());
-        response.put("count", pageList.getTotalCount());
-        return response.toString();
     }
 }
